+++ src/three/objects/scene.ts
@@ -1,19 +1,51 @@
 import * as THREE from "three";
+import { crosshair } from "./crosshair";
 
-const _scene = new THREE.Scene();
+export const renderer = new THREE.WebGLRenderer({ antialias: true });
+export const camera = new THREE.PerspectiveCamera(75, window.innerWidth / window.innerHeight, 0.1, 1000);
+
+export const scene = new THREE.Scene();
+
 const floorGeometry = new THREE.PlaneBufferGeometry(1000, 1000, 100, 100);
 const floorMaterial = new THREE.MeshBasicMaterial( { 
     color: 0x565656
 } );
-const _floor = new THREE.Mesh(floorGeometry, floorMaterial);
 
-_scene.background = new THREE.Color( 0x878787 );
-_scene.fog = new THREE.Fog( 0x878787, 0, 750 );
-_scene.add(new THREE.AxesHelper(100));
+export const floor = new THREE.Mesh(floorGeometry, floorMaterial);
+
+scene.background = new THREE.Color( 0x878787 );
+scene.fog = new THREE.Fog( 0x878787, 0, 750 );
+scene.add(new THREE.AxesHelper(100));
 
 
-_floor.rotateX(-Math.PI / 2);
-_scene.add(_floor);
+floor.rotateX(-Math.PI / 2);
+scene.add(floor);
 
-export const scene = _scene;
-export const floor = _floor;
+export const toStart = () => {
+    camera.position.set(-10, 10, -10);
+    camera.rotation.set(0, 180 * Math.PI / 180, 0); // Turn aroundas
+};
+
+toStart();
+
+camera.aspect = window.innerWidth / window.innerHeight;
+camera.updateProjectionMatrix();
+    
+renderer.setPixelRatio( window.devicePixelRatio );
+renderer.setSize( window.innerWidth, window.innerHeight );
+
+// Crosshair
+camera.add(crosshair);
+
+window.addEventListener("resize", () => {
+    camera.aspect = window.innerWidth / window.innerHeight;
+    camera.updateProjectionMatrix();
+
+    renderer.setSize( window.innerWidth, window.innerHeight );
+    render();
+}, false);
+
+
+const render = () => {
+    renderer.render(scene, camera);
+};
